09-06 19:34 Member 테이블 BLOB 컬럼 추가
