+++ src/main/java/Mingle/MingleProject/controller/GatheringController.java
@@ -1,6 +1,7 @@
 package Mingle.MingleProject.controller;
 
 import Mingle.MingleProject.dto.GatheringDTO;
+import Mingle.MingleProject.dto.MemberDTO;
 import Mingle.MingleProject.service.CityService;
 import Mingle.MingleProject.service.GatheringService;
 import Mingle.MingleProject.service.MemberService;
@@ -29,9 +30,12 @@ public class GatheringController {
         GatheringDTO gatheringDTO = gatheringService.findByGathering(id);
         model.addAttribute("GatheringHome", gatheringDTO);
 
-        List<MemberDTO> gatheringMemberDTO = gatheringService.findByGatheringMember(gatheringDTO.getGName());
+        List<MemberDTO> gatheringMemberDTO = memberService.findByGatheringMember(gatheringDTO.getGName());
         System.out.println("gatheringMemberDTO = " + gatheringMemberDTO);
         model.addAttribute("GatheringMember", gatheringMemberDTO);
+
+        int gatheringHeadcount = memberService.findByGatheringHeadcount(gatheringDTO.getGName());
+        model.addAttribute("headcount", gatheringHeadcount);
 
         return "Gathering_Home";
     }
