+++ src/main/java/Mingle/MingleProject/repository/MemberRepository.java
@@ -22,21 +22,30 @@ public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
     @Query("SELECT m.mId FROM MemberEntity m WHERE m.mName = ?1 AND m.mEmail = ?2")
     String findMemberIdByNameAndEmail(String mName, String mEmail);
 
+    // 사용자 정의 JPQL 쿼리를 사용하여 mEmail과 mId 같은 레코드를 조회
+    @Query("SELECT m FROM MemberEntity m WHERE m.mEmail = ?1 AND m.mId = ?2")
+    List<MemberEntity> findMembersByEmailAndId(String mEmail, String mId);
+
+    // mId와 mEmail을 기반으로 mPwd를 조회하는 메서드
+    @Query("SELECT m.mPwd FROM MemberEntity m WHERE m.mId = ?1 AND m.mEmail = ?2")
+    String findMemberPwdByIdAndEmail(String mId, String mEmail);
+
+    // 모임에 포함된 멤버를 조회하는 메서드
+    @Query(value="select m FROM MemberEntity m WHERE m.mGGathering like %:gatheringName%")
+    List<MemberEntity> findByGatheringMember(@Param("gatheringName") String gatheringName);
+
     //UPDATE MEMBER SET M_INTRODUCTION = 'test' WHERE M_ID = 'himedia1'
     /*@Query(value = "UPDATE MEMBER SET M_INTRODUCTION = 'test' WHERE M_ID = 'himedia1'", nativeQuery = true)*/
 
-    @Modifying
-    @Query(value = "UPDATE MemberEntity m SET m.mIntroduction  = :mIntroduction where m.mId =:mId")
-    void updateMIntroduction(@Param("mIntroduction") String mIntroduction, @Param("mId") String mId);
-
 
     @Modifying
-    @Query(value = "UPDATE MemberEntity m SET m.mPiProfileimg  = :mPiProfileimg where m.mId ='himedia'")
-    void updateMIntroduction(@Param("mPiProfileimg") String mPiProfileimg);
+    @Query(value = "UPDATE MemberEntity m SET m.mIntroduction  = :mIntroduction where m.mId ='himedia'")
+    void updateMIntroduction(@Param("mIntroduction") String mIntroduction);
 
-    // 모임에 포함된 멤버를 조회하는 메서드
-    @Query(value="select m FROM MemberEntity m WHERE m.mGGathering = :gatheringName")
-    List<MemberEntity> findByGatheringMember(@Param("gatheringName") String gatheringName);
+//    @Modifying
+//    @Query(value = "UPDATE MemberEntity m SET m.mPiProfileimg  = :mPiProfileimg where m.mId ='himedia'")
+//    void updateMPiProfileimg(@Param("mPiProfileimg") String mPiProfileimg);
 
-    void updateMPiProfileimg(@Param("mPiProfileimg") String mPiProfileimg);
+    @Query(value = "select COUNT(*) from MemberEntity m where m.mGGathering like %:gName%")
+    int findByGatheringHeadcount(@Param("gName") String gName);
 }
